Use lover-cased and dash-separated file names
