--- scripts/Repositories/IItem.ts
@@ -1,5 +1,0 @@
-export interface IItem {
-  ID: string
-  Name: string
-  Parent: string
-}
