+++ scripts/components/content.tsx
