+++ scripts/components/footer.tsx
