+++ scripts/repositories/interfaces.ts
@@ -1,6 +1,10 @@
 'use strict'
 
-import { IItem } from 'IItem'
+export interface IItem {
+  ID: string
+  Name: string
+  Parent: string
+}
 
 export interface IRepository {
   getChildren(id: string): Promise<IItem[]>
